Build tutorial devcontainer image from ubuntu:focal base
Instead of using the 9.8 GB universal codespace image, build a custom dapr
image for tutorials to reduce the image size.

This iteration is 6.56 GB, which likely lacks some requirements for all of
the quickstart examples and requires further additions. For comparison, the
Dockerfile-dev image produces an image that's 7.89 GB.

Note also that this change results in a slower Codespace startup time, as
building a multistep fresh image is slower than downloading a larger image,
so there are no immediate benefits until the smaller image is published to
the daprio registry.